+++ src/storage/store_key.rs
@@ -53,9 +53,10 @@ impl StoreKey {
     /// Validates a key according to the following rule from the specification:
     /// - a key is a Unicode string, where the final character is not a `/` character.
     /// Additionally, a key which starts with '/' is invalid even though this is not explicit in the specification.
+    /// A key cannot be an empty string.
     #[must_use]
     pub fn validate(key: &str) -> bool {
-        !key.starts_with('/') && !key.ends_with('/')
+        !key.starts_with('/') && !key.ends_with('/') && !key.eq("")
     }
 
     /// Returns true if the key has prefix `prefix`.
@@ -70,19 +71,15 @@ impl StoreKey {
         StorePrefix::new(self.0.clone() + "/").unwrap_or_else(|_| StorePrefix::root())
     }
 
-    /// Returns the parent of this key, or [`None`] this key is the root key and it has no parent.
+    /// Returns the parent of this key.
     #[must_use]
-    pub fn parent(&self) -> Option<StorePrefix> {
-        if self.0.is_empty() {
-            None
-        } else {
-            let key_split: Vec<_> = self.as_str().split('/').collect();
-            let mut parent = key_split[..key_split.len() - 1].join("/");
-            if !parent.is_empty() {
-                parent.push('/');
-            }
-            Some(unsafe { StorePrefix::new_unchecked(&parent) })
+    pub fn parent(&self) -> StorePrefix {
+        let key_split: Vec<_> = self.as_str().split('/').collect();
+        let mut parent = key_split[..key_split.len() - 1].join("/");
+        if !parent.is_empty() {
+            parent.push('/');
         }
+        unsafe { StorePrefix::new_unchecked(&parent) }
     }
 }
 
@@ -99,5 +96,29 @@ impl From<&StorePrefix> for StoreKey {
         let prefix = prefix.as_str();
         let key = prefix.strip_suffix('/').unwrap_or(prefix);
         unsafe { Self::new_unchecked(key.to_string()) }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn store_prefix() {
+        assert!(StoreKey::new("a").is_ok());
+        assert!(StoreKey::new("a/").is_err());
+        assert!(StoreKey::new("/a").is_err());
+        assert_eq!(
+            StoreKey::new("a").unwrap().to_prefix(),
+            StorePrefix::new("a/").unwrap()
+        );
+        assert_eq!(
+            StoreKey::new("a/b").unwrap().parent(),
+            StorePrefix::new("a/").unwrap()
+        );
+        assert_eq!(
+            StoreKey::new("a").unwrap().parent(),
+            StorePrefix::new("").unwrap()
+        );
     }
 }